merchant_app/requirements.txt
fastapi
alembic
pandas
pydantic
python-dotenv
SQLAlchemy
uvicorn
xlsxwriter
pymysql
cryptography
redis